Welkomst bericht bij inloggen, navigatiebalk opmaak.

+++ assets/js/pages/profiel_wijzigen.js
@@ -8,11 +8,11 @@ $( document ).ready(function() {
         document.getElementById("firstName").value = data[0].voornaam;
         document.getElementById("lastName").value = data[0].achternaam;
 
-        /* Controleert welk gegeven er in de database staat en markeert daarna de juiste radio button optie */
+        /* Controleert welk gender er in de database staat en markeert daarna de juiste radio button optie */
 
-        if(data[0].gender === "man"){
+        if(data[0].gender === "Man"){
         $(':input:radio:eq(0)').attr('checked', true);
-        }else if(data[0].gender === "vrouw"){
+        }else if(data[0].gender === "Vrouw"){
             $(':input:radio:eq(1)').attr('checked', true);
         } else {
             $(':input:radio:eq(2)').attr('checked', true);
@@ -23,87 +23,90 @@ $( document ).ready(function() {
         document.getElementById("geboortedatum").value = data[0].datum;
         */
 
-        /* Controlleert welke woonplaats in de database staat en geeft daarna de juiste optie weer */
+        /* Controleert welke woonplaats in de database staat en geeft daarna de juiste optie van het dropwdown menu weer in het formulier */
         switch (data[0].woonplaats){
-            case "Drenthe" :
+            case "drenthe" :
                 $('#woonplaats').find('option:eq(1)').attr('selected', true);
                 break;
-            case "Flevoland" :
+            case "flevoland" :
                 $('#woonplaats').find('option:eq(2)').attr('selected', true);
                 break;
-            case "Gelderland" :
+            case "friesland" :
                 $('#woonplaats').find('option:eq(3)').attr('selected', true);
                 break;
-            case "Groningen" :
+            case "gelderland" :
                 $('#woonplaats').find('option:eq(4)').attr('selected', true);
                 break;
-            case "Limburg" :
+            case "groningen" :
                 $('#woonplaats').find('option:eq(5)').attr('selected', true);
                 break;
-            case "Noord-Brabant" :
+            case "limburg" :
                 $('#woonplaats').find('option:eq(6)').attr('selected', true);
                 break;
-            case "Noord-Holland" :
+            case "noordbrabant" :
                 $('#woonplaats').find('option:eq(7)').attr('selected', true);
-                break
-            case "Overijssel" :
-                $('#woonplaats').find('option:eq(8)').attr('selected', true);
                 break;
-            case "Utrecht" :
+            case "noordholland" :
+                $('#woonplaats').find('option:eq(8)').attr('selected', true);
+                break
+            case "overijssel" :
                 $('#woonplaats').find('option:eq(9)').attr('selected', true);
                 break;
-            case "Zeeland" :
+            case "utrecht" :
+                $('#woonplaats').find('option:eq(10)').attr('selected', true);
+                break;
+            case "zeeland" :
                 $('#woonplaats').find('option:eq(11)').attr('selected', true);
                 break;
-            case "Zuid-Holland" :
+            case "zuidholland" :
                 $('#woonplaats').find('option:eq(12)').attr('selected', true);
                 break;
         }
 
         document.getElementById("budget").value = data[0].budget;
 
-        /* Controlleert welke woonplaats in de database staat en geeft daarna de juiste optie weer */
+        /* Controleert welke woonplaats in de database staat en geeft daarna de juiste optie van het dropdown menu weer in het formulier*/
         switch (data[0].reisbestemming){
-            case "Nederland" :
+            case "nederland" :
                 $('#reisbestemming').find('option:eq(1)').attr('selected', true);
                 break
-            case "Bonaire" :
+            case "bonaire" :
                 $('#reisbestemming').find('option:eq(2)').attr('selected', true);
                 break;
-            case "Bulgarije" :
+            case "bulgarije" :
                 $('#reisbestemming').find('option:eq(3)').attr('selected', true);
                 break;
-            case "Curaçao" :
+            case "curacao" :
                 $('#reisbestemming').find('option:eq(4)').attr('selected', true);
                 break;
-            case "Egypte" :
+            case "egypte" :
                 $('#reisbestemming').find('option:eq(5)').attr('selected', true);
                 break;
-            case "Gambia" :
+            case "gambia" :
                 $('#reisbestemming').find('option:eq(6)').attr('selected', true);
                 break;
-            case "Griekenland" :
+            case "griekenland" :
                 $('#reisbestemming').find('option:eq(7)').attr('selected', true);
                 break;
-            case "Indonesië" :
+            case "indonesie" :
                 $('#reisbestemming').find('option:eq(8)').attr('selected', true);
                 break;
-            case "Italië" :
+            case "italie" :
                 $('#reisbestemming').find('option:eq(9)').attr('selected', true);
                 break;
-            case "Kaapverdië" :
+            case "kaapverdie" :
                 $('#reisbestemming').find('option:eq(10)').attr('selected', true);
                 break;
-            case "Macedonie" :
+            case "macedonie" :
                 $('#reisbestemming').find('option:eq(11)').attr('selected', true);
                 break;
-            case "Portugal" :
+            case "portugal" :
                 $('#reisbestemming').find('option:eq(12)').attr('selected', true);
                 break;
-            case "Spanje" :
+            case "spanje" :
                 $('#reisbestemming').find('option:eq(13)').attr('selected', true);
                 break;
-            case "Turkije" :
+            case "turkije" :
                 $('#reisbestemming').find('option:eq(14)').attr('selected', true);
                 break;
         }
@@ -113,7 +116,7 @@ $( document ).ready(function() {
     });
 
 
-    /* Gegevens opslaan */
+    /* Gewijzigde gegevens opslaan en naar database sturen */
     $("#opslaan").on("click", function () {
 
         let firstName = document.getElementById("firstName").value;
@@ -151,7 +154,7 @@ $( document ).ready(function() {
     });
 
 
-    /* Preview van foto */
+    /* Preview van profielfoto */
     $(function () {
         $("#profilePicture").on("change", function () {
             FYSCloud.Utils
@@ -166,7 +169,6 @@ $( document ).ready(function() {
         });
 
         /* Profielfoto uploaden */
-
         $("#fileUploadButton").on("click", function () {
             let userid =  FYSCloud.Session.get("userid");
                 FYSCloud.Utils
@@ -197,21 +199,27 @@ $( document ).ready(function() {
         });
 
         /* Account verwijderen */
+        $(function() {
+        $('#verwijder').click(function(e) {
+            if(confirm("Are you sure?"))
+            {
+                FYSCloud.API.queryDatabase(
+                    "DELETE FROM profiel WHERE id = ?", [FYSCloud.Session.get("userid")]
+                ).done(function(data) {
+                    console.log(data);
+                    alert("Account is succesvol verwijderd.")
+                    window.location.href = "index.html";
+                }).fail(function(reason) {
+                    console.log(reason);
+                    alert("Account verwijderen mislukt.")
+                });
+            }
+            else
+            {
+                e.preventDefault();
+            }
+                });
+            });
 
-        $("#verwijder").on("click", function () {
-
-        FYSCloud.API.queryDatabase(
-            "DELETE FROM profiel WHERE id = ?", [FYSCloud.Session.get("userid")]
-        ).done(function(data) {
-            console.log(data);
-            alert("account is succesvol verwijderd.")
-            window.location.href = "index.html";
-        }).fail(function(reason) {
-            console.log(reason);
-            alert("Account verwijderen mislukt.")
-        });
         });
     });
-});
-
-
